requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
filelock
fsspec
gitdb
GitPython
huggingface-hub
idna
Jinja2
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging
pandas
pillow
protobuf
pyarrow
pydeck
Pygments
python-dateutil
pytz
PyYAML
referencing
regex
requests
rich
rpds-py
safetensors
six
smmap
streamlit
sympy
tenacity
tokenizers
toml
toolz
torch
tornado
tqdm
transformers
triton
typing_extensions
tzdata
urllib3
watchdog
